make package names plural too
