+++ src/main/java/com/google/cloud/tools/app/impl/cloudsdk/CloudSdkAppEngineInstances.java
@@ -17,8 +17,8 @@
 package com.google.cloud.tools.app.impl.cloudsdk;
 
 import com.google.cloud.tools.app.api.AppEngineException;
-import com.google.cloud.tools.app.api.instance.AppEngineInstances;
-import com.google.cloud.tools.app.api.instance.InstancesSelectionConfiguration;
+import com.google.cloud.tools.app.api.instances.AppEngineInstances;
+import com.google.cloud.tools.app.api.instances.InstancesSelectionConfiguration;
 import com.google.cloud.tools.app.impl.cloudsdk.internal.process.ProcessRunnerException;
 import com.google.cloud.tools.app.impl.cloudsdk.internal.sdk.CloudSdk;
 import com.google.cloud.tools.app.impl.cloudsdk.util.Args;
