+++ src/main/java/com/google/cloud/tools/app/impl/cloudsdk/CloudSdkAppEngineVersions.java
@@ -17,9 +17,9 @@
 package com.google.cloud.tools.app.impl.cloudsdk;
 
 import com.google.cloud.tools.app.api.AppEngineException;
-import com.google.cloud.tools.app.api.version.AppEngineVersions;
-import com.google.cloud.tools.app.api.version.VersionsListConfiguration;
-import com.google.cloud.tools.app.api.version.VersionsSelectionConfiguration;
+import com.google.cloud.tools.app.api.versions.AppEngineVersions;
+import com.google.cloud.tools.app.api.versions.VersionsListConfiguration;
+import com.google.cloud.tools.app.api.versions.VersionsSelectionConfiguration;
 import com.google.cloud.tools.app.impl.cloudsdk.internal.process.ProcessRunnerException;
 import com.google.cloud.tools.app.impl.cloudsdk.internal.sdk.CloudSdk;
 import com.google.cloud.tools.app.impl.cloudsdk.util.Args;
